% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_admonition_to_table.R
\name{convert_admonition_to_table}
\alias{convert_admonition_to_table}
\title{Convert the text of an admonition to
a regular markdown table}
\usage{
convert_admonition_to_table(text)
}
\arguments{
\item{text}{a character vector}
}
\value{
a text with regular markdown
}
\description{
Convert the text of an admonition to
a regular markdown table
}
