% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_files_to_markdown.R
\name{convert_files_to_markdown}
\alias{convert_files_to_markdown}
\title{Convert the text of multiple MkDocs files into one
markdown file, adding pagebreaks in between}
\usage{
convert_files_to_markdown(mkdocs_filenames, md_filename)
}
\arguments{
\item{mkdocs_filenames}{path to one or more files that
contain MkDocs markdown}

\item{md_filename}{path to the file that will be created
and will have markdown}
}
\value{
Nothing
}
\description{
Convert the text of multiple MkDocs files into one
markdown file, adding pagebreaks in between
}
