% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_text_to_markdown.R
\name{convert_text_tags_to_markdown}
\alias{convert_text_tags_to_markdown}
\title{Convert a text with MkDocs markdown extensions
to regular markdown}
\usage{
convert_text_tags_to_markdown(text)
}
\arguments{
\item{text}{a character vector}
}
\value{
a text with regular markdown
}
\description{
Convert a text with MkDocs markdown extensions
to regular markdown
}
