% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admonition_type_emoji.R
\name{get_admonition_type_emoji}
\alias{get_admonition_type_emoji}
\title{Get the emoji corresponding to the admonition}
\usage{
get_admonition_type_emoji(admonition_type)
}
\arguments{
\item{admonition_type}{the type of admonition}
}
\value{
an emoji, as text
}
\description{
Get the emoji corresponding to the admonition
}
\examples{
get_admonition_type_emoji("warning")
}
