% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_md_table_text.R
\name{is_md_table_text}
\alias{is_md_table_text}
\title{Detects if a text is a markdown table}
\usage{
is_md_table_text(text)
}
\arguments{
\item{text}{a text, as character vector}
}
\description{
Detects if a text is a markdown table
}
